NO longer generate the class dynamicly, just 2 different classes for 2 supported versions will be enough

+++ src/com/cnaude/chairs/sitaddons/ChairEffects.java
@@ -50,7 +50,7 @@ public class ChairEffects {
                 for (Player p : Bukkit.getOnlinePlayers()) {
                     if (plugin.getPlayerSitData().isSitting(p)) {
                         if (p.hasPermission("chairs.sit.health")) {
-                            double pHealthPcnt = ((double) p.getHealth()) / (double) p.getMaxHealth() * 100d;
+                            double pHealthPcnt = (p.getHealth()) / p.getMaxHealth() * 100d;
                             if ((pHealthPcnt < plugin.sitMaxHealth) && (p.getHealth() < p.getMaxHealth())) {
                                 double newHealth = plugin.sitHealthPerInterval + p.getHealth();
                                 if (newHealth > p.getMaxHealth()) {
